requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
flask-cors==3.0.10
mysql-connector-python==8.1.0
bcrypt==4.0.1
requests==2.31.0
Werkzeug==2.3.7
